requirements.txt
# Core Dependencies
requests
httpx>=0.27.0  # Async HTTP client for MiDaS MCP server
python-dotenv
pydantic>=2.0.0
python-multipart

# Auth & Security (JWT)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# OpenAI Agents SDK
openai
openai-agents

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# RAG & Vector Store
chromadb>=0.4.18
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Computer Vision (MiDaS)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL (optional, can use SQLite for dev)
aiosqlite>=0.19.0  # Async SQLite database driver

# Text Processing & Classification
sentence-transformers>=2.2.0  # For hybrid routing classifier
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
python-json-logger>=2.0.0

# Web Search
tavily-python>=0.3.0  # Web search and retrieval for RAG